refactor wave outcome reviews router
